Improve database part

+++ src/database/post_tag.ts
@@ -1,63 +1,62 @@
+import { snakeCase } from "change-case";
 import { PostTag } from "../entity/post_tag.js";
-import { getClient } from "./postgresql.js";
+import { getClient } from "../postgresql.js";
 
 interface PostTagRepository {
-  findByPost(postId: string): Promise<PostTag[]>;
-  findByTag(tagId: string): Promise<PostTag[]>;
-  bind(postId: string, tagId: string): Promise<void>;
-  unbind(postId: string, tagId: string): Promise<void>;
+  find(postTag: Partial<PostTag>): Promise<PostTag[]>;
+  create(postId: string, tagId: string): Promise<void>;
+  remove(postId: string, tagId: string): Promise<void>;
 }
 
 class PostTagRepositoryImpl implements PostTagRepository {
-  async findByPost(postId: string): Promise<PostTag[]> {
+  async find(postTag: Partial<PostTag>): Promise<PostTag[]> {
+    if (Object.entries(postTag).length === 0) {
+      throw new Error("Empty postTag");
+    }
+
     const client = await getClient();
-    const result = await client.query(
-      "SELECT * FROM post_tag WHERE post_id = $1",
-      [postId],
-    );
+
+    const conditions: string[] = [];
+    const values: any[] = [];
+
+    for (const [key, value] of Object.entries(postTag)) {
+      conditions.push(`${snakeCase(key)} = $${values.length + 1}`);
+      values.push(value);
+    }
+
+    const query = `SELECT * FROM post_tag WHERE ${conditions.join(" AND ")}`;
+    const result = await client.query(query, values);
     client.release();
-    return result.rows.map((row) => {
-      const postTag: PostTag = {
-        postId: row.post_id,
-        tagId: row.tag_id,
-      };
-      return postTag;
-    });
+
+    return result.rows.map((row) => ({
+      postId: row.post_id,
+      tagId: row.tag_id,
+    }));
   }
 
-  async findByTag(tagId: string): Promise<PostTag[]> {
-    const client = await getClient();
-    const result = await client.query(
-      "SELECT * FROM post_tag WHERE tag_id = $1",
-      [tagId],
-    );
-    client.release();
-    return result.rows.map((row) => {
-      const postTag: PostTag = {
-        postId: row.post_id,
-        tagId: row.tag_id,
-      };
-      return postTag;
-    });
-  }
-
-  async bind(postId: string, tagId: string): Promise<void> {
+  async create(postId: string, tagId: string): Promise<void> {
+    if (postId.length === 0 || tagId.length === 0) {
+      throw new Error("Empty postId or tagId");
+    }
     const client = await getClient();
     await client.query(
       "INSERT INTO post_tag (post_id, tag_id) VALUES ($1, $2)",
-      [postId, tagId],
+      [postId, tagId]
     );
     client.release();
   }
 
-  async unbind(postId: string, tagId: string): Promise<void> {
+  async remove(postId: string, tagId: string): Promise<void> {
+    if (postId.length === 0 || tagId.length === 0) {
+      throw new Error("Empty postId or tagId");
+    }
     const client = await getClient();
     await client.query(
       "DELETE FROM post_tag WHERE post_id = $1 AND tag_id = $2",
-      [postId, tagId],
+      [postId, tagId]
     );
     client.release();
   }
 }
 
-export { PostTagRepository };
+export { PostTagRepository, PostTagRepositoryImpl };
