+++ src/database/tag.ts
@@ -1,6 +1,6 @@
 import { snakeCase } from "change-case";
 import { Tag } from "../entity/tag.js";
-import { getClient } from "./postgresql.js";
+import { getClient } from "../postgresql.js";
 
 interface TagRepository {
   find(id: string): Promise<Tag | null>;
@@ -12,6 +12,9 @@ interface TagRepository {
 
 class TagRepositoryImpl implements TagRepository {
   async find(id: string): Promise<Tag | null> {
+    if (id.length === 0) {
+      throw new Error("Empty id");
+    }
     const client = await getClient();
     const result = await client.query("SELECT * FROM tag WHERE id = $1", [id]);
     client.release();
@@ -39,18 +42,36 @@ class TagRepositoryImpl implements TagRepository {
     });
   }
 
-  async update(id: string, tag: Partial<Omit<Tag, "id">>): Promise<void> {
-    const client = await getClient();
-    for (const [key, value] of Object.entries(tag)) {
-      await client.query(
-        `UPDATE tag SET ${snakeCase(key)} = $1 WHERE id = $2`,
-        [value, id]
-      );
+  async update(id: string, tag: Partial<Tag>): Promise<void> {
+    if (id.length === 0) {
+      throw new Error("Empty id");
     }
+    const conditions = [];
+    const values = [];
+    for (const [key, value] of Object.entries(tag)) {
+      if (key === "id") {
+        continue;
+      }
+      conditions.push(`${snakeCase(key)} = $${conditions.length + 1}`);
+      values.push(value);
+    }
+    if (conditions.length === 0) {
+      throw new Error("Empty tag");
+    }
+    const client = await getClient();
+    await client.query(
+      `UPDATE tag SET ${conditions.join(", ")} WHERE id = $${
+        conditions.length + 1
+      }`,
+      [...values, id]
+    );
     client.release();
   }
 
   async create(tag: Tag): Promise<void> {
+    if (tag.id.length === 0 || tag.name.length === 0) {
+      throw new Error("Empty id or name");
+    }
     const client = await getClient();
     await client.query("INSERT INTO tag (id, name) VALUES ($1, $2)", [
       tag.id,
@@ -60,10 +81,13 @@ class TagRepositoryImpl implements TagRepository {
   }
 
   async remove(id: string): Promise<void> {
+    if (id.length === 0) {
+      throw new Error("Empty id");
+    }
     const client = await getClient();
     await client.query("DELETE FROM tag WHERE id = $1", [id]);
     client.release();
   }
 }
 
-export { TagRepository };
+export { TagRepository, TagRepositoryImpl };
